# OpenAPI/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: User Management API
  version: 1.0.0
  description: API for managing users with CRUD operations.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/users:
    get:
      summary: Get all users
      tags: [User]
      operationId: getAllUser
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      summary: Create a user
      tags: [User]
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/users/{id}:
    get:
      summary: Get user by ID
      tags: [User]
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
    put:
      summary: Update user
      tags: [User]
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      summary: Delete user
      tags: [User]
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted

components:
  schemas:
    UserRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string 
          example: John Doe
        email:
          type: string
          example: john.doe@example.com